LLM-Python-Django/requirements.txt
django>=5.1,<6
djangorestframework>=3.15,<4
django-cors-headers>=4.4,<5
openai>=1.30,<2
python-dotenv>=1.0,<2
gunicorn>=22.0,<24

